Parse validation configuration from markdown. Still needs lots of caching and things to reduce repeat work.
